feat: 로또여러개를 생성하는 util 클래스

+++ src/main/java/lotto/controller/LottoOutputController.java
@@ -52,7 +52,7 @@ public class LottoOutputController {
         lottoResultOutputView.printDashLine();
     }
 
-    private void printTotalProfitRate(int amount, Map<LottoRank, Long> lottoRankCountMap) {
+    private void printTotalProfitRate(long amount, Map<LottoRank, Long> lottoRankCountMap) {
         lottoResultOutputView.printTotalProfitRate(amount, lottoRankCountMap);
     }
 }
